Defer visible popover mutations outside SwiftUI updates
