+++ Packages/macOS/CmuxAppKitSupportUI/Sources/CmuxAppKitSupportUI/Popover/ArrowlessPopoverRootViewUpdatePolicy.swift
@@ -1,5 +1,13 @@
 struct ArrowlessPopoverRootViewUpdatePolicy {
-    static func shouldUpdateRootView(isPresented: Bool, popoverIsShown: Bool) -> Bool {
-        isPresented || popoverIsShown
+    enum Strategy {
+        case none
+        case immediate
+        case deferredVisible
+    }
+
+    static func rootViewUpdateStrategy(isPresented: Bool, popoverIsShown: Bool) -> Strategy {
+        if popoverIsShown { return .deferredVisible }
+        if isPresented { return .immediate }
+        return .none
     }
 }
