+++ cmuxTests/MobileHostAuthorizationTests.swift
@@ -1,4 +1,5 @@
 import CMUXMobileCore
+import CmuxIrohTransport
 import Foundation
 @preconcurrency import Network
 import Testing
@@ -1171,7 +1172,6 @@ actor TestMobileHostIndependentEventWriter: MobileHostIndependentEventWriting {
 }
 struct ImmediateMobileHostIrohClock: CmxIrohRelayClock {
     private let instant = Date(timeIntervalSince1970: 1_700_000_000)
-
     func now() -> Date { instant }
     func sleep(until _: Date) async throws {}
 }
